Update gifts index page.
