+++ Sources/Views/PricingV2.swift
@@ -120,118 +120,11 @@ public struct PricingV2: HTML {
     }
 
     WhatToExpect()
-    PricingFAQ()
+    FAQModule(faqs: .allFaqs)
     WhatPeopleAreSaying()
     Companies()
     if currentUser == nil {
       GetStartedModule(style: .gradient)
-    }
-  }
-}
-
-private struct Lane<PriceDetails: HTML, Features: HTML, CallToAction: HTML>: HTML {
-  var name: String
-  var annualPricePerMonth: Dollars<Int>?
-  var priceDetails: PriceDetails?
-  var features: Features
-  var callToAction: CallToAction
-
-  init(
-    _ name: String,
-    annualPricePerMonth: Dollars<Int>? = nil,
-    @HTMLBuilder priceDetails: () -> PriceDetails = { Never?.none },
-    @HTMLBuilder features: () -> Features,
-    @HTMLBuilder callToAction: () -> CallToAction
-  ) {
-    self.name = name
-    self.annualPricePerMonth = annualPricePerMonth
-    self.priceDetails = priceDetails()
-    self.features = features()
-    self.callToAction = callToAction()
-  }
-
-  var body: some HTML {
-    Card {
-      VStack {
-        Header(4) { HTMLText(name) }
-          .inlineStyle("margin-top", "1.5rem")
-        if let annualPricePerMonth {
-          HStack(alignment: .center) {
-            div {
-              Header(3) { HTMLText("$\(annualPricePerMonth)") }
-                .inlineStyle("font-size", "2rem")
-                .inlineStyle("font-weight", "300")
-            }
-
-            if let priceDetails {
-              div {
-                priceDetails
-              }
-              .inlineStyle("font-size", "0.75rem")
-            }
-          }
-        }
-        ul {
-          features
-        }
-        .color(.gray500.dark(.gray800))
-        .linkColor(.gray150.dark(.gray900))
-        .flexContainer(direction: "column", rowGap: "0.5rem")
-        .listStyle(.reset)
-        .inlineStyle("font-size", "0.875rem")
-        .inlineStyle("margin", "0 0 1.5rem")
-      }
-      .color(.black.dark(.white))
-    } footer: {
-      callToAction
-        .flexContainer(justification: "center")
-    }
-  }
-}
-
-private struct EpisodesStats {
-  var allEpisodes: Int = 0
-  var allHours: Int = 0
-  var freeEpisodes: Int = 0
-
-  init() {
-    @Dependency(\.episodes) var episodes
-    @Dependency(\.envVars.emergencyMode) var emergencyMode
-    @Dependency(\.date.now) var now
-
-    var allSeconds = 0
-    for episode in episodes() {
-      guard episode.publishedAt < now else { continue }
-      allEpisodes += 1
-      allSeconds += episode.length.rawValue
-      if !episode.isSubscriberOnly(currentDate: now, emergencyMode: emergencyMode) {
-        freeEpisodes += 1
-      }
-    }
-    allHours = allSeconds / 3600
-  }
-}
-
-private struct PricingFAQ: HTML {
-  var body: some HTML {
-    PageModule(title: "FAQ", theme: .content) {
-      VStack(spacing: 3) {
-        HTMLForEach([Faq].allFaqs) { faq in
-          div {
-            Header(4) {
-              HTMLText(faq.question)
-            }
-            .color(.black.dark(.offWhite))
-
-            HTMLMarkdown(faq.answer)
-          }
-        }
-      }
-      .color(.gray300.dark(.gray850))
-      .linkColor(.gray150.dark(.gray900))
-      .linkUnderline(true)
-      .inlineStyle("margin", "0 auto")
-      .inlineStyle("width", "60%", media: .desktop)
     }
   }
 }
